week-1-api-security/day-05-graphql-attack-detector/pyproject.toml
[tool.poetry]
name = "day-05-graphql-attack-detector"
version = "1.0.0"
description = "GraphQL introspection, depth attack, batching abuse, field suggestion oracle, and injection detector"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python              = "^3.11"
httpx               = "^0.27"
fastapi             = "^0.110"
uvicorn             = {extras = ["standard"], version = "^0.27"}
strawberry-graphql  = {extras = ["fastapi"], version = "^0.219"}

[tool.poetry.dev-dependencies]
pytest     = "^8.1"
pytest-cov = "^5.0"

[tool.pytest.ini_options]
testpaths     = ["gql_scanner/tests"]
log_cli       = true
log_cli_level = "INFO"

[tool.coverage.run]
source = ["gql_scanner"]
omit   = ["*/tests/*", "*/targets/*"]

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
